add create board form validation and dispatch action
