+++ src/index.tsx
@@ -5,7 +5,6 @@ import './index.css'
 import App from './App'
 import ThemeContextProvider from './contexts/ThemeContext'
 import SidePanelProvider from './contexts/SidePanelContext'
-import DataContextProvider from './contexts/DataContext'
 import OverlayContextProvider from './contexts/OverlayContext'
 import StateManagement from './contexts/StateManagement'
 
@@ -13,15 +12,13 @@ const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement)
 root.render(
   <React.StrictMode>
     <ThemeContextProvider>
-      <DataContextProvider>
-        <OverlayContextProvider>
-          <SidePanelProvider>
-            <StateManagement>
-              <App />
-            </StateManagement>
-          </SidePanelProvider>
-        </OverlayContextProvider>
-      </DataContextProvider>
+      <OverlayContextProvider>
+        <SidePanelProvider>
+          <StateManagement>
+            <App />
+          </StateManagement>
+        </SidePanelProvider>
+      </OverlayContextProvider>
     </ThemeContextProvider>
   </React.StrictMode>
 )
